fix: more phpstan suggested fixes.

+++ app/Console/Commands/SubmissionsAutoProcess/FetchCASNumbersAuto.php
@@ -279,7 +279,7 @@ class FetchCASNumbersAuto extends Command
                 )) {
                     $delay = $baseDelay * $attempt + (rand(0, 500) / 1000);
                     Log::info("Retrying after {$delay}s due to connection issue...");
-                    usleep($delay * 1000000);
+                    usleep((int) ($delay * 1000000));
 
                     continue;
                 }
@@ -287,7 +287,7 @@ class FetchCASNumbersAuto extends Command
 
             if ($attempt < $maxRetries) {
                 $delay = $baseDelay * $attempt;
-                usleep($delay * 1000000);
+                usleep((int) ($delay * 1000000));
             }
         }
